requirements.txt
scrapegraphai
ollama
playwright
psycopg2-binary
python-dotenv
pytest
pytest-cov
pytest-asyncio
